Fix rekey epoch replay-window leak that permanently stalls at key_epoch=256
Root-caused a real incident: a long-running xenia-peer session (any
transport/codec) permanently froze its entire send loop right after logging
"session rekey acknowledged key_epoch=256" -- the process stayed alive
(background heartbeat ticker kept logging) but never sent or received
another frame. Discovered live during unrelated Android file-transfer
testing; reproducible on any sufficiently long session since rekeys happen
roughly every 4 frames under the default RekeyPolicy.

Root cause: `Session::current_key_epoch` (a purely internal, never-
transmitted counter used to key per-epoch replay-window state, not to be
confused with the per-session `epoch: u8` byte actually embedded in the
wire nonce) was a `u8`, incremented via `wrapping_add` on every
`install_key`. The real bug enabling the wraparound to actually matter was
upstream of the wraparound itself: `install_key` tracked only a single
`prev_session_key`/`prev_key_epoch`/`prev_key_expires_at` slot, unconditionally
overwritten on every subsequent rekey -- so a rekey cadence faster than
`DEFAULT_REKEY_GRACE` (5s) meant `tick()`'s `drop_epoch` almost never
actually ran in steady state, and every historical epoch's replay-window
entries leaked until the 256-value space wrapped and collided with the
stale epoch-0 entries from session start. The next rekey proposal (an
AEAD-sealed control-lane envelope, itself starting over at epoch=0/seq=0
after the wrap) then got rejected by the receiver's replay window as an
exact duplicate of that stale entry, `Session::open` returned
`WireError::OpenFailed`, and both peers silently stopped making progress:
the viewer's frame-open loop just `warn!`s and `continue`s forever, while
the host's rekey handshake blocks forever on `rekey_ack_rx.recv().await`
waiting for an ack that will never arrive.

Two-part fix:

1. Replace the single `prev_session_key`/`prev_key_epoch`/`prev_key_expires_at`
   slot with `prev_keys: Vec<PrevKey>`, each tracking its own key + epoch +
   expiry. `install_key` now pushes rather than overwrites; `tick()` drains
   every entry whose own grace period has passed (via `drop_epoch` per
   entry) instead of clobbering all-but-the-latest before it ever expires.
   This is the fix that actually matters at any realistic rekey cadence --
   it also fixes a related correctness gap where an envelope genuinely
   in-flight across two or more rapid rekeys would have been rejected
   outright by the old single-slot design.
2. Widen `current_key_epoch`/`PrevKey::epoch`/`ReplayWindow`'s `key_epoch`
   dimension from `u8` to `u32` as defense in depth, so a future collision
   would require billions of rekeys instead of 256. Purely internal
   bookkeeping -- never serialized, so this touches no wire format, no
   nonce byte layout, and no existing test-vector byte assertions (checked
   against `test_vector_validation.rs`/`violation_vectors.rs`).

Verified: all 90 existing tests (xenia-wire) and 71+ (xenia-peer-core, the
main downstream consumer) pass unchanged; added two regression tests
(`rekey_past_256_epochs_with_tick_never_stalls` drives 300 rekeys at a
realistic fast cadence with `tick()` interleaved and asserts every frame
keeps opening -- this is exactly the freeze if it regresses --
`envelope_from_two_rekeys_back_opens_within_its_own_grace_window` proves
the Vec<PrevKey> correctness improvement directly) plus a replay_window.rs
unit test proving epoch 256 no longer collides with epoch 0.

+++ fuzz/fuzz_targets/fuzz_replay_window.rs
@@ -21,7 +21,11 @@ fuzz_target!(|data: &[u8]| {
         }
         let source_id = u64::from_le_bytes(chunk[0..8].try_into().unwrap());
         let payload_type = chunk[8];
-        let key_epoch = chunk[9];
+        // `key_epoch` is `u32` (widened from `u8`); only fuzzing the
+        // u8 sub-range is intentional here -- the interesting bitmap/
+        // shift arithmetic this target exercises is orthogonal to the
+        // epoch dimension's own value range.
+        let key_epoch = chunk[9] as u32;
         let seq = u64::from_le_bytes(chunk[10..18].try_into().unwrap());
         let _ = w.accept(source_id, payload_type, key_epoch, seq);
     }
